Add "throw" statements to the traceback, make SEE_Print*() more flexible.
Extended traceback helps when a throw occurs in the toplevel of a script.
Allows you to track down the statement that caused the exception.

The ctxt traceback field is now also zeroed during SEE_TRY() to avoid 
having garbage when printing a traceback.

Null FILEs can now be passed to the SEE_Print*() functions, and they
are converted to standard error output.  This is handy to know when
inside gdb.