Add --with-longjmperror option, and correct the way --with-native-dtoa works
The longjmperror() handler may be defined by an application to override SEE's
but it may not be clear if that is desirable, so I've made it an option.

Also, when I cleaned up the arg enabling code, I noticed that the native-dtoa
code may never be enabled. So, now --enable-native-dtoa=yes doesn't check
for the presence of functions in libc: it simply assumes they're there.